feat: 流水线新增构建矩阵 #4518

+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineVMBuildService.kt
@@ -241,7 +241,7 @@ class PipelineVMBuildService @Autowired(required = false) constructor(
                 )
                 containerBuildDetailService.containerStarted(
                     buildId = buildId,
-                    containerId = vmSeqId.toInt(),
+                    containerId = vmSeqId,
                     containerBuildStatus = buildStatus
                 )
             }
