+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/vmbuild/EngineVMBuildService.kt
@@ -268,7 +268,7 @@ class EngineVMBuildService @Autowired(required = false) constructor(
                 )
                 containerBuildDetailService.containerStarted(
                     buildId = buildId,
-                    containerId = vmSeqId.toInt(),
+                    containerId = vmSeqId,
                     containerBuildStatus = buildStatus
                 )
             }
